added a local version of metad_analysis, would be good to have an option to select whether things are run on a cluster or locally. Essentially just specifying whether or not to run with mpirun or just use a GMX executable
