+++ setup.py
@@ -73,6 +73,7 @@ setup(
             'parameterize_foldamer = terphenyl_simulations.scripts:parameterize_foldamer',
             'average_rtt = terphenyl_simulations.scripts:calculate_average_rtt',
             'REMD_setup = terphenyl_simulations.scripts:REMD_setup',
+            'mw_metad_analysis = terphenyl_simulations.analysis_workflows.mw_metad:main',
             'metad_analysis = terphenyl_simulations.analysis_workflows.metad:main'
         ]
     }
